refactor : #32 공통 위젯 생성 후 구현
- 계정과 프로필 위젯을 공통위젯으로 파일을 따로 만들어 구현
- 계정 : account_widget.dart, 프로필 : profile_widget.dart
